+++ src/Qualys/functions/public/Remove-QualysUser.ps1
@@ -5,8 +5,6 @@
     Adds a new user to Qualys
 .PARAMETER Credential
     This API call only supports basic HTTP authentication. You must provide your credentials separately for this function.
-.PARAMETER ExternalID
-    Set a custom External ID (required for SSO)
 .PARAMETER SendEmail
     Specifies whether the new user will receive an email notification with a secure link to their login credentials.
 .PARAMETER Role