+++ src/Qualys/functions/public/Set-QualysUser.ps1
@@ -5,14 +5,10 @@
     Adds a new user to Qualys
 .PARAMETER Credential
     This API call only supports basic HTTP authentication. You must provide your credentials separately for this function.
+.PARAMETER Login
+    Specifies the Qualys user login of the user account you wish to edit.
 .PARAMETER ExternalID
     Set a custom External ID (required for SSO)
-.PARAMETER SendEmail
-    Specifies whether the new user will receive an email notification with a secure link to their login credentials.
-.PARAMETER Role
-    Specifies the user role. A valid value is: manager, unit_manager, scanner, reader, contact or administrator.
-.PARAMETER BusinessUnit
-    Specifies the user’s business unit.
 .PARAMETER AssetGroups
     Specifies the asset groups assigned to the user, when theuser role is Scanner, Reader or Contact.
 .PARAMETER FirstName
@@ -34,44 +30,36 @@
 .PARAMETER State
     Specifies the user’s state.
 .EXAMPLE
-    $NewUserSplat = @{
+    Set-QualysUser -Credential $Credential -Login testuser -AssetGroups TestGroup
+.EXAMPLE
+    $SetUserSplat = @{
+        Login = testuser
         Phone = '555-555-555'
         Address = 'University of Illinois'
         city = 'Urbana'
         country = 'United States of America'
         state = 'Illinois'
-        Role = 'scanner'
         FirstName = 'Jane'
         LastName = 'Doe'
         Title = 'Test User'
         email = 'JaneDoe@test.null'
         Credential = $Credential
     }
-    Add-QualysUser @NewUserSplat
+    Set-QualysUser @SetUserSplat
 #>
-function Add-QualysUser{
+function Set-QualysUser{
     [CmdletBinding()]
     param (
-        [Switch]$SendEmail,
         [Parameter(Mandatory=$true)]
-        [String]$Role,
-        [String]$BusinessUnit = 'Unassigned',
+        [String]$Login,
         [String[]]$AssetGroups,
-        [Parameter(Mandatory=$true)]
         [String]$FirstName,
-        [Parameter(Mandatory=$true)]
         [String]$LastName,
-        [Parameter(Mandatory=$true)]
         [String]$Title,
-        [Parameter(Mandatory=$true)]
         [String]$Email,
-        [Parameter(Mandatory=$true)]
         [String]$Phone,
-        [Parameter(Mandatory=$true)]
         [String]$Address,
-        [Parameter(Mandatory=$true)]
         [String]$City,
-        [Parameter(Mandatory=$true)]
         [String]$Country,
         [String]$State,
         [String]$ExternalID,
@@ -81,30 +69,24 @@ function Add-QualysUser{
 
     process{
 
+
+
         $RestSplat = @{
             Method = 'POST'
             RelativeURI = 'msp/user.php'
             Credential = $Credential
             Body = @{
-                action = 'add'
-                send_email = [string][int]$SendEmail.IsPresent
-                business_unit = $BusinessUnit
-                user_role = $Role
-                first_name = $FirstName
-                last_name = $LastName
-                title = $Title
-                phone = $Phone
-                email = $Email
-                address1 = $Address
-                city = $City
-                country = $Country
-                state = $State
-                external_id = $ExternalID
+                action = 'edit'
             }
         }
 
+        [String[]]$Exclusions = ('Credential', 'AssetGroups')
+        $PSBoundParameters.Keys | Where-Object -FilterScript {($_ -notin $Exclusions) -and $_} | ForEach-Object -Process {
+            $RestSplat.Body.$_ = $PSBoundParameters[$_]
+        }
+
         If($AssetGroups){
-            $RestSplat.Body['asset_groups'] = (($AssetGroups).Trim() -join ", ")
+            $RestSplat.Body['asset_groups'] = (($AssetGroups).Trim() -join ",")
         }
 
         $Response = Invoke-QualysRestCall @RestSplat